Cleaned MafIO.py. No longer a link.
